requirements.txt
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.31
yfinance>=0.2.40
feedparser>=6.0
pandas>=2.0
numpy>=1.24
SQLAlchemy>=2.0
transformers>=4.40
torch>=2.0
streamlit>=1.30
plotly>=5.20
